refactor file strcuture
